Bugfix: change clamps to follow domain scale
Radial and Line clamps work with angle or parameter from 0 to 1
but clamps are moved according to junction.delta which works with
cell sizes.

Clamps are now changed so that angle is translated to actual movement
(bigger radii-bigger junction.delta and v.v.) and lines' parameter
is actual displacement instead of interval 0...1.