Drop sketch optimization, add a custom example
Sketch optimization by 'minimizing quad energy' does not work in
real-life as it is too unstable numerically and slow.

RoundSquare sketch works because it is convex and smoothing is
sufficient.

Also smoothing is now done by averaging distances to neighbour corners
AND quad centers. Not sure if it helps.

Sketches with concave corners will not profit from smoothing.
Now to solve that; options are numerous, one more difficult than the
other.