Add an Array class
Curves (and possibly other future entities) that are defined by a large
number of points create a huge bottleneck when transforming.
Each Point is transformed separately and within that every
transformation matrix must be constructed separately and new and new
numpy arrays are created which takes a lot of time.

Array holds the same numpy array with *all* the points and transforms
everything in one go. Improvement on a gear case: 45s > 17s (the whole
case, not just the transforms).