Rework all of chopping and propagation
Edge chop is thrown out because it burdens the user with
micromanagement. Instead, the 'take' parameter is introduced back into
Chop.

Grading and Chops are now friends and grading specification got smarter.

Propagation in block_list is now much simpler.

The Loft example with edge grading produced bad quality cells so it was
changed.