Fix bug with rollback during optimization
When a vertex is far from given curve/surface, a Clamp will try to find
the appropriate parameters to put that vertex onto that geometry.
This may fail or even when not, might result in a bad mesh. During
optimization rollback was performed simply by using the original clamp
parameters; in the above case, the 'mesh worsening' happened even before
the optimization of the vertex and rollback did not undo the changes.

Now rollback is done using actual vertex position instead of clamp
parameters.